Refactored list species code and removed local file check on the push task
